Refactored data and analysis packages into additional modules.
