+++ sci_analysis/__init__.py
@@ -2,7 +2,6 @@
 Sub packages:
     data - sci_analysis data types
     analysis - sci_analysis test and calculation classes and functions
-    operations - sci_analysis general functions
     graphs - graphing classes
 
 The analysis and operations modules are loaded.
@@ -14,10 +13,11 @@ clean and strip.
 """
 from __future__ import absolute_import
 
-from sci_analysis.analysis import analysis
-from sci_analysis.graphs import graph
-from sci_analysis.data import data
-__all__ = ["data", "analysis", "graph", "operations"]
-from .analysis.analysis import Comparison, NormTest, TTest, LinearRegression,\
-    Correlation, Anova, Kruskal, EqualVariance, VectorStatistics, GroupStatistics,\
-    analyze
+# from sci_analysis.analysis import analysis
+# from sci_analysis.graphs import graph
+# from sci_analysis.data import data
+__all__ = ["data", "analysis", "graph"]
+# from analysis import analyze
+# from .analysis.analysis import Comparison, NormTest, TTest, LinearRegression,\
+#     Correlation, Anova, Kruskal, EqualVariance, VectorStatistics, GroupStatistics,\
+#     analyze
