+++ sci_analysis/graphs/graph.py
@@ -21,9 +21,7 @@ from numpy import polyfit, polyval, sort, arange, array, linspace, mgrid, vstack
 from scipy.stats import probplot, gaussian_kde
 
 # local imports
-# from ..data.data import assign
-from sci_analysis.data import Vector
-from sci_analysis.operations.data_operations import is_dict, is_group
+from data import Vector, is_dict, is_group
 # TODO: Add preferences back in a future version
 # from ..preferences.preferences import GraphPreferences
 # from six.moves import range
