--- sci_analysis/operations/__init__.py
@@ -1,5 +1,0 @@
-"""sci_analysis package: operations
-Modules:
-    data_operations - functions for working with data types
-"""
-# from data_operations import *
