+++ sci_analysis/test/test_equalvariance.py
@@ -2,7 +2,7 @@ import unittest
 import numpy as np
 import scipy.stats as st
 
-from sci_analysis.analysis.analysis import EqualVariance, MinimumSizeError, NoDataError
+from analysis import EqualVariance, MinimumSizeError, NoDataError
 
 
 class MyTestCase(unittest.TestCase):
