+++ sci_analysis/test/test_kruskal.py
@@ -2,7 +2,7 @@ import unittest
 import numpy as np
 import scipy.stats as st
 
-from ..analysis.analysis import Kruskal, MinimumSizeError, NoDataError
+from analysis import Kruskal, MinimumSizeError, NoDataError
 
 
 class MyTestCase(unittest.TestCase):
