+++ sci_analysis/test/test_linregress.py
@@ -2,8 +2,8 @@ import unittest
 import numpy as np
 import scipy.stats as st
 
-from ..analysis.analysis import LinearRegression, MinimumSizeError, NoDataError
-from ..data.data import UnequalVectorLengthError
+from analysis import LinearRegression, MinimumSizeError, NoDataError
+from data import UnequalVectorLengthError
 
 
 class MyTestCase(unittest.TestCase):
