+++ sci_analysis/test/test_statistics.py
@@ -2,7 +2,7 @@ import unittest
 import numpy as np
 import scipy.stats as st
 
-from ..analysis.analysis import VectorStatistics, MinimumSizeError, NoDataError
+from analysis import VectorStatistics, MinimumSizeError, NoDataError
 
 
 class MyTestCase(unittest.TestCase):
