+++ sci_analysis/test/test_twosampleks.py
@@ -2,7 +2,7 @@ import unittest
 import numpy as np
 import scipy.stats as st
 
-from ..analysis.analysis import TwoSampleKSTest, MinimumSizeError, NoDataError
+from analysis import TwoSampleKSTest, MinimumSizeError, NoDataError
 
 
 class TestTwoSampleKS(unittest.TestCase):
